ai/crawler/requirements.txt
selenium
sqlalchemy
psycopg2-binary
python-dotenv